Add files via upload
This project involves Hotel Data Analysis with SQL and Power BI, including building a database, analyzing data, connecting Power BI to the database, and visualizing data with interactive charts and tables. It's beginner-friendly, follows best practices in data analysis, and provides a valuable addition to a data analyst's portfolio. Open-source tools are used to complete the project, with deliverables including a database, SQL queries, and Power BI reports.